Start migrating `DecorateDiagCompat::Builtin` items to `DecorateDiagCompat::Dynamic`
