+++ compiler/rustc_parse/src/parser/expr.rs
@@ -19,10 +19,9 @@ use rustc_ast::{
     StmtKind, Ty, TyKind, UnOp, UnsafeBinderCastKind, YieldKind,
 };
 use rustc_data_structures::stack::ensure_sufficient_stack;
-use rustc_errors::{Applicability, Diag, PResult, StashKey, Subdiagnostic};
+use rustc_errors::{Applicability, Diag, Diagnostic, PResult, StashKey, Subdiagnostic};
 use rustc_literal_escaper::unescape_char;
 use rustc_session::errors::{ExprParenthesesNeeded, report_lit_error};
-use rustc_session::lint::BuiltinLintDiag;
 use rustc_session::lint::builtin::BREAK_WITH_LABEL_AND_LOOP;
 use rustc_span::edition::Edition;
 use rustc_span::{BytePos, ErrorGuaranteed, Ident, Pos, Span, Spanned, Symbol, kw, respan, sym};
@@ -1921,11 +1920,20 @@ impl<'a> Parser<'a> {
                         _ => false,
                     }
                 {
+                    let span = expr.span;
                     self.psess.buffer_lint(
                         BREAK_WITH_LABEL_AND_LOOP,
                         lo.to(expr.span),
                         ast::CRATE_NODE_ID,
-                        BuiltinLintDiag::BreakWithLabelAndLoop(expr.span),
+                        rustc_errors::DynamicDiag::new(move |dcx, level| {
+                            errors::BreakWithLabelAndLoop {
+                                sub: errors::BreakWithLabelAndLoopSub {
+                                    left: span.shrink_to_lo(),
+                                    right: span.shrink_to_hi(),
+                                },
+                            }
+                            .into_diag(dcx, level)
+                        }),
                     );
                 }
